feat: Surface sizing API, cross-platform SDK alignment, memory & rendering fixes
- Introduce ISurfaceSizeProvider and SurfaceManager.setSurfaceSizeProvider
  so hosts can synchronously inject surface size and feed it into Yoga;
  Surface now accepts px by default and converts to a2ui units internally
- updateSurfaceSize now suppresses no-op size notifications; batched
  component-change messages are delivered in order
- ComponentFactory exposes a custom-component measure hook; async
  measurement interface is hardened against bad input
- Fix Tabs text color sourcing, ChoicePicker multi-select attribute,
  Android font-weight parsing, HarmonyOS Lottie invisible, Android
  Gallery scroll jumping
- Android Tabs now parses tab-font-color / tab-font-color-selected
- Android & iOS playgrounds support streamed UpdateComponents
  rendering with a sync/async toggle
- Add Android ProGuard rules for external deps and reflection-accessed
  classes; add macro config to the iOS podspec

+++ platforms/android/src/main/java/com/amap/agenui/render/style/StyleHelper.java
@@ -140,7 +140,7 @@ public class StyleHelper {
         if ("bold".equals(value)) return true;
         if ("normal".equals(value)) return false;
         try {
-            return Integer.parseInt(value) >= 500;
+            return Double.parseDouble(value) >= 500;
         } catch (NumberFormatException ignored) {
             return false;
         }
